refactor: extract clef config to utils/clef.ts
Moves clef derivation logic from constants.ts to dedicated utils/clef.ts:
- Keeps constants.ts clean with only static data
- Re-exports CLEF_CONFIG, ClefConfig, getClefConfig for backward compatibility
- All derivation logic (CLEF_REFERENCES, deriveClefConfig, helpers) in clef.ts

See ADR-007: Open-Closed Clef Reference Pattern
